pyproject.toml
[project]
name = "zmag"
version = "0.0.13"
description = "Craft APIs with ZeroMQ and GraphQL"
authors = [
    { name = "hlop3z", email = "23062270+hlop3z@users.noreply.github.com" },
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
dependencies = ["pyzmq>=26.2.0", "orjson>=3.10.7"] # "zmag"

[project.optional-dependencies]
debug = ["uvicorn[standard]>=0.30.6", "starlette>=0.38.2"]
server = ["spoc>=0.1.35", "strawberry-graphql>=0.237.3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.377",
    "mypy>=1.11.1",
    "isort>=5.13.2",
    "black>=24.8.0",
    "pylint>=3.2.6",
    "ruff>=0.6.2",
    "bandit>=1.7.9",
    "watchdog>=4.0.2",
    "ssort>=0.13.0",
]

[tool.uv.sources]
# zmag = { workspace = true }
